+++ Assets/Wilson/Scripts/RandomSpawner.cs
@@ -2,8 +2,11 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class randomSpawner : MonoBehaviour
+public class RandomSpawner : MonoBehaviour
 {
+    public Transform[] spawnPoints;
+    public GameObject[] enemyPrefabs;
+
     // Start is called before the first frame update
     void Start()
     {
@@ -13,6 +16,12 @@ public class randomSpawner : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        
+        if (Input.GetMouseButtonDown(0))
+        {
+            int randEnemy = Random.Range(0, enemyPrefabs.Length);
+            int randSpawPoint = Random.Range(0, spawnPoints.Length);
+
+            Instantiate(enemyPrefabs[0], spawnPoints[randSpawPoint].position, transform.rotation);
+        }
     }
 }
