--- Assets/Wilson/randomSpawner.cs
@@ -1,27 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class RandomSpawner : MonoBehaviour
-{
-    public Transform[] spawnPoints;
-    public GameObject[] enemyPrefabs;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        if (Input.GetMouseButtonDown(0))
-        {
-            int randEnemy = Random.Range(0, enemyPrefabs.Length);
-            int randSpawPoint = Random.Range(0, spawnPoints.Length);
-
-            Instantiate(enemyPrefabs[0], spawnPoints[randSpawPoint].position, transform.rotation);
-        }
-    }
-}
