--- Assets/Wilson/randomSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 631baadf7f392ef47af976468384b9c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
